Fix build: test classes also separated by package
