+++ src/test/java/org/jpatterns/gof/behavioral/CommandTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.behavioral;
 
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
