+++ src/test/java/org/jpatterns/gof/behavioral/StrategyTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.behavioral;
 
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
