+++ src/test/java/org/jpatterns/gof/behavioral/TemplateMethodTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.behavioral;
 
 import static org.easymock.EasyMock.*;
 import static org.junit.Assert.assertEquals;
