+++ src/test/java/org/jpatterns/gof/creational/AbstractFactoryTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.creational;
 
 import org.junit.Test;
