+++ src/test/java/org/jpatterns/gof/creational/SingletonEagerTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.creational;
 
 import static junit.framework.Assert.assertNotNull;
 import org.junit.Test;
