+++ src/test/java/org/jpatterns/gof/structural/CompositeTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.structural;
 
 import static junit.framework.Assert.assertEquals;
 import org.junit.Test;
