+++ src/test/java/org/jpatterns/gof/structural/FlyweightTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.structural;
 
 import org.junit.Test;
