+++ src/test/java/org/jpatterns/gof/structural/ProxyTest.java
@@ -1,4 +1,4 @@
-package org.jpatterns.gof;
+package org.jpatterns.gof.structural;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
